Added examples on how to use DataManagement for injecting of test data
Fixed overall structure:
1. IMDBPage is only for a Page and not for components
2. PageObject means a component in the page.

SubComponents should not extend IMDBPage since it is not a page but a component within a Page.
If a IMBD Page always has a navigation header, subcomponents do not have that.
Incorrect modelling leads to defects in the future.

Currently due to incorrect structure, the Factory is unable to initialize the PAgeObject due to recursive.